requirements.txt
pandas
datetime
google.cloud
gcsfs
matplotlib
requests
python-dotenv